Start things off with a bang.
This is one helluva commit. I'm a bad person.

This basically starts us off with a fully-fledged datastore interface.
We have an in-memory implementation and a Postgres implementation. We
have test coverage for a lot of the interface, and we have migrations
working so tests can be run in parallel without stepping all over each
other--we just stand up a new database for each test. That probably
won't be sustainable forever, but that sounds like a problem for Future
Paddy.

We've vendored everything. The only thing off here is that I'm using a
doctored version of the migration helper; for whatever reason, it was
using a global for the database connection, so trying to migrate
multiple databases at once made it cry. So I fixed that. You can see the
commit here: https://github.com/impractical/migrate/commit/bd05d6207339ba9942e8c2973d959fb18eede898

This also sets up a Travis file, which will (hopefully) run the tests.
Or at least the non-Postgres ones. Let's walk before running.

We even have a license! The only things I can think of that are missing and
seriously wrong are:

* README
* code of conduct
* contributing guide

And let's face it, Rome wasn't built in a day. Also, I need this damn
code off my computer before my harddrive gives out just to spite me.