Add an apiv1 package.
Add a simple apiv1 package that exposes the logic from the tokens
package over an HTTP interface.

The apiv1 package gets its own struct versions of RefreshToken and
RefreshTokenChange so that we can try to keep it backwards compatible,
and so we can isolate the JSON encoding needs between API response and
anything else that may want to encode things in JSON (e.g., caching).

The main way this package will be used is to call GetRouter, passing the
base path (`/` if you're not serving from a subdirectory). That will
return an http.Handler that can then be passed to http.Handle.