Add a tokensd daemon that serves the API.
Add a binary (tokensd) that will serve the API endpoints from apiv1.
we're basically just bootstrapping a context (using Postgres, in this
case), then setting up the API handlers, and serving.

We also added Version and Hash variables to the tokens package, to hold
information from Git. Version will hold the tag and Hash will hold the
git hash. A build.sh script is included to generate the binary and
inject these values at build time. The tokensd binary reports them on
startup.

We also added tokensd/tokensd to the .gitignore, because there's really
no reason for us to commit a binary file to git.