Lay out the basis of our client.
Add a client package that introduces a Manager interface. This Manager
interface is how consumers will interact with the tokens service. We
offer two built-in implementations: an in-memory Manager, for use in
stubbing and testing calls to the tokens service, and an HTTP API
Manager that actually calls the tokens service, for use in production
environments.

There's still a ways to go on this. It hasn't actually been tested
against anything just yet, and we're going to want to add some more
niceties to it--a way to track which consumer is making which requests,
a way to turn on verbose logging to see the API calls actually being
made, etc.--but this is enough of a start that it makes sense to get it
off my harddrive and onto the internet somewhere.