Get Dockerfile working.
I had to tinker with the Dockerfile a bit to get it working properly.
The big difference now is our build.sh script has to copy the SQL files
into the tokensd directory so it can include the migrations in the
docker image, so it can run migrations at startup. I wish I could just
include the SQL migrations from the parent directory, but them's the
breaks.

The Dockerfile has a bunch of cosmetic changes that don't really do
anything. The bug that was screwing me up was the fact that `go build`
was generating dynamically linked libraries instead of statically linked
libraries, so they couldn't be run in `busybox`. To fix this,
`CGO_ENABLED` is now an environment variable that `build.sh` will pass
to the `go build` command it runs. Just run `CGO_ENABLED=0 ./build.sh`
and you'll get a docker-ready binary.

We also needed to include ca-certificates.crt so we could do SSL
resolution from within the binary. I just copied what ships with Fedora
23.