Drop Datastore support, update .travis.yml
No longer support Datastore as a backing Storer. Just not worth it.

While we're in there, update .travis.yml. Namely, run the tests for all
our packages, and generate a combined coverage report for the entire
project, as well as the packages themselves. This is handled by our
coverage.sh script. We also turned on linting. It won't fail the build,
but it will be logged.