Hash our RefreshTokens at rest.
As RefreshTokens are technically long-term credentials, we should really
be encrypting them at rest. RFC 6819 Section 5.1.4.1.3 clearly
identifies refresh tokens as something that should be encrypted at
rest[1].

To accomplish this, we made the following changes:

* RefreshToken IDs are no longer 32 bytes of cryptographically random
  data hex-encoded. They are now UUIDs in string format, though they
  don't necessarily have to be. So... they lost 16 bytes of crypto
  random data, and gained some hyphens, basically.
* RefreshTokens now have a Value property. This is the _actual_ value of
  the RefreshToken that should be exposed to users (more on this below).
  Importantly, this is never stored anywhere.
* RefreshTokens now have a Hash property. This is the cryptographically
  secure hash generated from the Value property, and is what we actually
  store in the database to compare against.
* RefreshTokens now have a HashSalt property, which is the salt that was
  used to generate their Hash property.
* RefreshTokens now have a HashIterations property, which is the number
  of iterations used to calculate their Hash property.
* We now have another index of the combination of the token's Hash,
  HashSalt, and HashIterations properties, ensuring the combination is
  unique. This guarantees that we're not handing out a value that has
  already been used, which could lead to giving one user access to
  another's account. The odds are astronomical, but it's something we
  can trivially prevent, so let's.
* We have a new SQL migration to get the new columns and the new
  constraint into Postgres.
* We have a new test to ensure the new constraint is working. Well, it's
  actually just expanding an old test. I cheated a bit. Whatever.
* We have a new error, ErrTokenHashAlreadyExists, which is returned when
  the new constraint is violated.
* FillTokenDefaults now generates a Value and/or
  Hash/HashSalt/HashIterations value, as needed.
* We get GenerateTokenHash and GenerateTokenValue, which generate the
  hash and salt for a given value and number of iterations, and generate
  a random value, respectively.

As for what gets returned to the user... The Value is what they need,
it's what is replacing the old ID. But we have no way of generating the
Hash for a Value until we have already retrieved a RefreshToken from the
Storer. So we can't havea GetTokenByValue method. My proposed solution
to this is simple: when returning a RefreshToken to the user, return a
JWT of the ID and the Value. When they send it bakc, break out the two
pieces, retrieve the Token using its ID, and calculate the Hash from the
Value the user provided. If the Hashes match, it's a match.

[1]: https://tools.ietf.org/html/rfc6819#section-5.1.4.1.3