Add headers to the client.
Take care of our last TODO by adding some handy reporting headers to the
client. This includes things like the version of tokens that built the
client... (sort of, if the application including the client decides to
include that info), the hostname that issued the request, the
application that initiated the request (set when instantiating the
APIManager), etc.

Also, set the Content-Type, so our POST requests won't fail.