Change to an Impractical Labs repo, add Validate
Update our repo to the new Impractical Labs repo location, as this
didn't make sense as a Darling project.

Update our dependencies (trout and api), using the new Darling version
of trout.

Swap out our custom version and build script combo for the standardised
version package in darling.

Introduce the concept of validation for RefreshTokens. This is part of
the new idea that the tokens service will act as a grant provider in our
overarching auth architecture. Basically, other privileged services will
need to be able to verify that user-supplied strings are tokens, and are
valid. The idea is that the coordinating service will take the token as
a grant authorization, validate its authenticity, and generate the
requested grant. To be able to do that, we needed an API endpoint to
validate the token, and our clients needed to be able to interact with
it.

To help make some of the ideas in this clearer, we renamed the strings
being passed in used to _identify_ a token (e.g., in Get) to "id", and
use "token" as an opaque representation of the token's id and value in a
single string, as a user would present it.