Update to go1.7 and remove secondbit deps.
Update to go 1.7 and take advantage of the built-in context package, as
well as the new test hierarchies. This mostly involved rewriting the old
golang.org/x/net/context import paths to be just context. It also
involved updating the storer_test.go file to use our new test
hierarchies, meaning the update variation tests can finally run in
parallel. Not that that's a big deal for this repo, but it will be for
other repos.

Also, use this opportunity to purge our old dependencies (basically,
anything secondbit), replacing them with the updated versions.
code.secondbit.org/pqarrays.hg turned into code.impractical.co/pqarrays,
which is just a namespace move. github.com/secondbit/pan turned into
darlinggo.co/pan, which has a new API, which required some updates.
code.secondbit.org/uuid.hg was dropped, in favor of the package it was
wrapping, which has been updated to the point where the wrapper is
largely unnecessary now. This had minor API changes, but those were
dealt with rather easily.