Update to use latest darlinggo.co/api
Update to use the newest darlinggo.co/api, which uses slices of
errors instead of the hashicorp multierror library, which was
giving me problems. Now everything returns plain slices of errors,
which is clearer (this can return multiple errors) and has no such
import problems.

Also, we fixed the buildURL method of the APIManager, which was
ruining URLs by adding an extra slash as part of the protocol.
Oops.

Finally, the CreatedAt field on the Token is now a time.Time.