Fix API bugs
Make sure that every response gets wrapped inside the Response
object, which wasn't happening before due to laziness and mistakes.

Log a bunch of internal server errors, so we can at least see what
went wrong.

When updating a token, set the RefreshTokenChange's ID to the id
specified in the URL, to avoid invalid SQL generation.

Check the Used/Revoked properties when validating a token, so the
clients don't forget to and treat them as valid.