Use a JWT instead of a hashed value as our token.
Our refresh tokens used to be values, which were stored as salted hashes
with the original promptly thrown away. Now we don't generate that value
or store salted versions of it, we simply use a JWT with the token's ID.
We still need the database because refresh tokens are long-lived and
need to be revokable, but this allows us to forgo the expensive hashing.