Switch to new project layout.
Use our new project layout. This involves basically moving the storers
to a separate storers package and moving migrations to their own
package. This helps make test coverage clearer.

We also dropped the client and tokensd packages, as this is no longer
going to be a standalone binary, but part of a monolith when it's first
shipped.

Coverage reports now go in the .cover-reports folder, like all the other
projects, which means that folder needed to be added to the .gitignore.

Fixes #9.