Add datastore support.
Add support for the Google Cloud Datastore as a storer. This is
currently minorly buggy; we have a fundamental algorithmic error in
updating tokens, in which if we're trying to set Revoked to true on a
token or tokens that have Used set to true, Used will be set to false on
them, and vice versa. There's also something about too many tokens being
returned, which I don't really understand just yet.