Switch JWT signing method.
Update our NewPostgres helper to not return an error, that was never
gonna be used.

Add a ServiceID to our Dependencies, so we can fill that service ID in
for our Issuer on JWTs.

Change our JWTs to RS256 instead of HS256, and make sure to validate
that RS256 is being used when validating tokens.