Test all possible change filter combinations.
Replace all our filter-specific tests with a single beast of a test that
filters tokens to change by every conceivable combination of our filter
fields, and applies every possible change to them. Also test what
happens when no filter is specified, which should be an error. That
error had to be defined. Both our postgres and memory Storer
implementations needed to be updated to support the desired
functionality; memory was only able to filter on one field, and postgres
didn't throw an error when no filter was specified.