Switch to go-cmp, add AccountID to tokens.
Start tracking the AccountID on tokens, so when we refresh a token we
can fill in the AccountID on the grant, allowing us to easily trace the
provenance of tokens and revoke them if, say, an account is proven to be
compromised. To help in this specific scenario, this commit also adds
the ability to revoke tokens by their AccountID.

Also, switch from our hand-rolled token comparison code in tests to
using cmp.

🎵 Now Playing: Say You Love Me
🎵 Artist: Steve Grand
🎵 Album: All American Boy